🔨 Chore : Header 컴포넌트 분리하여 관리
- #3
- Header 컴포넌트 분리하여 가독성 개선 및 세분화
